requirements.txt
matplotlib
pydensecrf
numpy
Pillow
torch
torchvision
tqdm
cython